backend/requirements.txt
opencv-python
numpy
deepface
flask
pillow
flask-cors 